requirements.txt
python-telegram-bot==13.15 # telegram bot
requests # api
pytesseract # ocr
pillow # ocr
rapidfuzz # word matching pct (for ocr)